requirements.txt
fastapi
requests
uvicorn
sqlalchemy
python-dotenv
pydantic
psycopg2-binary